+++ projects/oblique/src/lib/autocomplete/_mocks/mock-highlight-text.module.ts
@@ -1,0 +1,8 @@
+import {NgModule} from '@angular/core';
+import {ObMockHighlightTextPipe} from './mock-highlight-text.pipe';
+
+@NgModule({
+	declarations: [ObMockHighlightTextPipe],
+	exports: [ObMockHighlightTextPipe]
+})
+export class ObMockHighlightTextModule {}