+++ projects/oblique/src/lib/autocomplete/_mocks/mock-highlight-text.pipe.ts
@@ -1,0 +1,13 @@
+import {Pipe, PipeTransform} from '@angular/core';
+import {DomSanitizer, SafeHtml} from '@angular/platform-browser';
+
+@Pipe({
+	name: 'obHighlightTextPipe'
+})
+export class ObMockHighlightTextPipe implements PipeTransform {
+	constructor(private readonly sanitizer: DomSanitizer) {}
+
+	transform(value: string, textToFind: string, cssClass = 'ob-searched-text', pattern = 'textToFind', regexFlags = 'gi'): SafeHtml {
+		return value;
+	}
+}