+++ projects/oblique/src/lib/autocomplete/_mocks/mock-option-label-icon.module.ts
@@ -1,0 +1,10 @@
+import {NgModule} from '@angular/core';
+import {ObMockOptionLabelIconDirective} from '../_mocks/mock-option-label-icon.directive';
+import {ObMockIconModule} from '../../icon/_mocks/mock-icon.module';
+
+@NgModule({
+	declarations: [ObMockOptionLabelIconDirective],
+	imports: [ObMockIconModule],
+	exports: [ObMockOptionLabelIconDirective]
+})
+export class ObMockOptionLabelIconModule {}