+++ projects/oblique/src/public_api.ts
@@ -174,4 +174,11 @@ export {obOutsideFilter} from './lib/global-events/outsideFilter';
 export {ObliqueModule} from './lib/oblique.module';
 export * from './lib/oblique-testing.module';
 export {ObMockMatElement, ObMockModule} from './lib/_mocks/mock-module';
+export {
+	ObAutocompleteComponent,
+	ObAutocompleteModule,
+	ObIAutocompleteInputOption,
+	ObIAutocompleteInputOptionGroup,
+	OptionLabelIconPosition
+} from './lib/autocomplete/autocomplete.module';
 export {ObPaginatorModule, ObPaginatorService} from './lib/paginator/ob-paginator.module';